Move all editing profile to their own nested url
