+++ src/app/dashboard/(tabs)/editProfile/layout.tsx
@@ -1,0 +1,15 @@
+import { ReactNode } from "react";
+
+export default async function EditProfileLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <div className="h-full">
+      <div className="max-w-2xl">
+        <>{children}</>
+      </div>
+    </div>
+  );
+}
